WebVTT Settings Reviewed Unit Tests (4th time)
Full:
http://kyle.barnhart.ca/2012/12/webvtt-settings-reviewed-unit-tests.html

The tests listed below do not include those that were already disabled.
These are tests the changed I made are causing them to fail. This is
likely due to the tests being new or that I made the tests stricter by
making sure the values were set properly (or default).